Extract toggle dark mode view func from controller file
